+++ src/components/ChefOfTheWeek/ChefOfTheWeek.scss
@@ -1,39 +1,80 @@
 @import "../../data/variables.scss";
 
 .chef-of-the-week-container {
-    display: flex;
-    flex-direction: column;
-    padding-top: 40px;
+  display: flex;
+  flex-direction: column;
+  padding-top: 10%;
+  @media (max-width: map-get($breakPoints, m)) {
+   align-items: flex-start;
+   margin-left: 5%;
+   margin-right: 5%;
+   padding-bottom: 5%;
+  }
 }
 
-.chef-of-the-week-title{
+.chef-of-the-week-title {
+  font-family: $font;
+  font-size: 30px;
+  font-weight: 200;
+  line-height: 35px;
+  letter-spacing: 1.25px;
+  text-align: center;
+  @media (max-width: map-get($breakPoints, m)) {
     font-family: $font;
-    font-size: 30px;
+    font-size: 18px;
     font-weight: 200;
     line-height: 35px;
     letter-spacing: 1.25px;
-    text-align: center;
+    text-align: left;
+   }
 }
 
 .img-text-container {
-    display: flex;
-    padding-top: 20px;
-    margin-left: 9%;
-    margin-right: 9%;
+  display: flex;
+  padding-top: 5%;
+  margin-left: 9%;
+  margin-right: 9%;
+  @media (max-width: map-get($breakPoints, m)) {
+    flex-direction: column;
+    margin-left: 0%;
+    margin-right: 0%;
+  }
 }
 
-.chef-img {
-    width: 100%;
-    height: 100%;
+.chef-img-desktop {
+  width: 100%;
+  height: 100%;
+  @media (max-width: map-get($breakPoints, m)) {
+    display: none;
+  }
+}
+
+.chef-img-mobile {
+  width: 100%;
+  height: 100%;
+  @media (min-width: map-get($breakPoints, m)) {
+    display: none;
+  }
 }
 
 .chef-of-the-week-text {
+  font-family: $font;
+  font-size: 24px;
+  font-weight: 200;
+  line-height: 35px;
+  letter-spacing: 1.25px;
+  text-align: justify;
+  padding-left: 4%;
+  padding-top: 2%;
+  @media (max-width: map-get($breakPoints, m)) {
     font-family: $font;
-    font-size: 24px;
+    font-size: 18px;
     font-weight: 200;
-    line-height: 35px;
+    line-height: 28px;
     letter-spacing: 1.25px;
     text-align: justify;
-    padding-left: 4%;
-    padding-top: 2%;
+    padding-left: 1%;
+    padding-right: 1%;
+    padding-top: 5%;
+  }
 }
