+++ src/components/ChefOfTheWeek/ChefOfTheWeek.tsx
@@ -4,18 +4,21 @@ const ChefOfTheWeek = ({
   title,
   chefName,
   text,
-  img,
+  desktopImg,
+  mobileImg,
 }: {
   title: string;
   chefName: string;
   text: string;
-  img: string;
+  desktopImg: string;
+  mobileImg: string;
 }) => {
   return (
     <div className="chef-of-the-week-container">
       <h1 className="chef-of-the-week-title">{title}</h1>
       <div className="img-text-container">
-        <img className='chef-img' src={img} alt={chefName}/>
+        <img className='chef-img-desktop' src={desktopImg} alt={chefName}/>
+        <img className='chef-img-mobile' src={mobileImg} alt={chefName}/>
         <p className="chef-of-the-week-text">{text}</p>
       </div>
     </div>
