+++ src/data/constants.ts
@@ -33,7 +33,8 @@ import kitchenMarketImg from "../assets/images/kitchenMarket.png";
 import mashyaImg from "../assets/images/mashya.png";
 
 // chef image
-import chefImg from "../assets/images/yossiShitrit.png";
+import desktopChefImg from "../assets/images/yossiShitrit.png";
+import mobileChefImg from "../assets/images/yossiShitritMobile.png";
 
 export const paths = {
   root: "/",
@@ -211,8 +212,9 @@ export const chefData = {
   title: "CHEF OF THE WEEK: ",
   text: "Chef Yossi Shitrit has been living and breathing his culinary dreams for more than two decades, including running the kitchen in his first restaurant, the fondly-remembered Violet, located in Moshav  Udim. Shitrit's creativity and culinary  acumen born of long experience  are expressed in the every detail of each and every dish.",
   chefName: "Yossi Shitrit",
-  chefImg: chefImg,
-  carouselTitle: "Yossi's Restaurants",
+  desktopImg: desktopChefImg,
+  mobileImg: mobileChefImg,
+  carouselTitle: "YOSSI'S RESTAURANTS",
 };
 
 export const chefRestaurantsData: CardData[] = [
@@ -242,33 +244,3 @@ export const chefRestaurantsData: CardData[] = [
   },
 ];
 
-export const carouselDynamicStyles = {
-  width: "60%",
-  height: "100%",
-  margin: "0 auto",
-};
-
-export const containerDynamicStyles = {
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "flex-start",
-  justifyContent: "flex-start",
-  rowGap: "30px",
-};
-
-export const containerRegDynamicStyles = {
-  margin: "0 auto",
-  paddingTop: "4%",
-  width: "100%",
-  display: "flex",
-  height: "100%",
-  flexDirection: "column",
-  justifyContent: "flex-start",
-  alignItems: "center",
-  rowGap: "30px",
-};
-
-if (window.matchMedia("(max-width: 768px)").matches) {
-  containerRegDynamicStyles.alignItems = "flex-start";
-  containerRegDynamicStyles.rowGap = "20px";
-}
