+++ src/pages/HomePage/HomePage.scss
@@ -48,7 +48,7 @@
   height: 100%;
   background-image: url("../../assets/icons/allRestaurantsButtonMobile.svg");
   background-size: contain;
-  @media (min-width: map-get($breakPoints, l)) {
+  @media (min-width: map-get($breakPoints, m)) {
     display: none;
   }
 }
